fix tpw_provider_fetch_assets_duration_seconds metrics
